22.1.26 webpack learning

+++ src/controllers/videoController.js
@@ -17,8 +17,12 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
   const id = req.params.id;
   const video = await Video.findById(id);
+  const { _id } = req.session.user;
   if (!video) {
     return res.status(404).render("404", { pageTitle: "video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -26,9 +30,13 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   const id = req.params.id;
   const { title, description, hashtags } = req.body;
+  const { _id } = req.session.user;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -69,6 +77,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.render("404", { pageTitle: "video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
